Move all dashboard provider stuff into a folder
