+++ ui/dashboards/src/context/DashboardProvider/index.ts
@@ -11,8 +11,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+export * from './DashboardAppSlice';
 export * from './DashboardProvider';
-export * from './DatasourceStoreProvider';
-export * from './QueryStringProvider';
-export * from './TemplateVariableProvider';
-export * from './TimeRangeProvider';
+export * from './LayoutsSlice';
